--- src/components/Category/ITissue.js
@@ -1,9 +1,0 @@
-import React from "react"
-
-import Posts from "../Postlist/PostData"
-
-function It() {
-  return <Posts pagename={"IT Issue"} id={"ITissue"} />
-}
-
-export default It
