--- src/components/Category/Project.js
@@ -1,9 +1,0 @@
-import React from "react"
-
-import Posts from "../Postlist/PostData"
-
-function Project() {
-  return <Posts pagename={"Project"} id={"Project"} />
-}
-
-export default Project
