--- src/components/Category/Session.js
@@ -1,9 +1,0 @@
-import React from "react"
-
-import Posts from "../Postlist/PostData"
-
-function Session() {
-  return <Posts pagename={"Session"} id={"Session"} />
-}
-
-export default Session
