--- src/components/Category/Study.js
@@ -1,9 +1,0 @@
-import React from "react"
-
-import Posts from "../Postlist/PostData"
-
-function Study() {
-  return <Posts pagename={"Study"} id={"Study"} />
-}
-
-export default Study
