+++ src/components/Postlist/PostList.js
@@ -1,28 +1,48 @@
 import React, { useState, useEffect } from "react"
+import axios from "axios"
+
 import PageName from "../Common/PageName"
-import Category from "../Common/Category"
-import countriesData from "./PostData"
 import Table from "./PostTable"
 
-function PostList() {
-  const [countries] = useState([...countriesData])
+function Posts() {
+  const [posts, setPosts] = useState(null)
+  const [loading, setLoading] = useState(false)
+  const [error, setError] = useState(null)
+
+  useEffect(() => {
+    const fetchPosts = async () => {
+      try {
+        // 요청 시작 : error 와 posts 를 초기화
+        setError(null)
+        setPosts(null)
+        // loading 상태 : true로
+        setLoading(true)
+        const response = await axios.get("https://jsonplaceholder.typicode.com/posts")
+        // "http://localhost:8080/${id}"
+        setPosts(response.data) // get data
+      } catch (e) {
+        setError(e)
+      }
+      setLoading(false)
+    }
+
+    fetchPosts()
+  }, [])
+
+  if (loading) return <div>로딩중</div>
+  if (error) return <div>에러</div>
+  if (!posts) return null
+
   return (
-    <>
-      <main>
-        <div className="container">
-          <PageName />
-          <div className="container" style={{ height: "auto", minHeight: "100%", paddingBottom: "20px" }}>
-            <div className="row">
-              <div className="col-3">
-                <Category />{" "}
-              </div>
-              <Table data={countries} rowsPerPage={10} />
-            </div>
-          </div>
-        </div>
-      </main>
-    </>
+    <div className="container">
+      <div className="row m-2 p-2 align-items-center">
+        <PageName pagename={window.location.pathname.substring(1)} />
+      </div>
+      <div className="row m-2">
+        <Table data={posts} rowsPerPage={10} />
+      </div>
+    </div>
   )
 }
 
-export default PostList
+export default Posts
